Place server requests in dedicated folder - Create services folder in frontend src directory - Rename services.js to posts-services.js and place it in services folder - Add auth-services.js, which will contain the server requests used in authorization process
